services: use common SetupDB function
Define and use a common SetupDB function for all services.
